requirements.txt
torch
torchvision
numpy
toml
wandb
tqdm
pillow
scipy
scikit-learn
albumentations==0.4.3
opencv-python
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.0.8
omegaconf==2.0.0
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
more-itertools>=8.0.0
transformers==4.3.1
taming-transformers
ptpt==0.0.25
x-transformers==0.22.0
rotary-embedding-torch
performer-pytorch
